Fix XFAIL behavior in conftest.py. (#88)
Fixes https://github.com/nod-ai/SHARK-TestSuite/issues/86. The
imperative `pytest.xfail()` does not actually XFAIL/XPASS, it acts as a
"skip". This switches to using `self.add_marker(pytest.mark.xfail(...))`
which actually works.

Now that XFAIL/XPASS work, new XFAIL sets were generated with
*
https://github.com/llvm/torch-mlir/commit/e48fe4588631e7a37a2899f9d4cd5c4cbc967481
*
https://github.com/openxla/iree/commit/e99179861c9e2af4796f51e09de271c885e88836

This also highlights a need for a way to automatically update the XFAIL
sets in the config files from a test run. I updated the config files
here manually and it was quite a chore :/