Add caching support for model testing (#170)
This commit adds caching support for model testing. This brings the
download time from ~20 mins to 23 seconds. With these changes, it is
feasible to add model testing as a presubmit in the iree repo which was
the reason for these changes.

Description of what's going on:  
First, we use an environment variable passed in (IREE_TEST_FILES) to
determine if we even want to use a cache or not (don't want to make it a
requirement). In CI, we set this because we want a persistent runner
cache. If this is set, we download files to the cache location specified
and symlink it to the testing directory. Otherwise, we just directly
download into the repo, and it won't be using any caching mechanism. The
test runner doesn't need to know about caching at all, and even
initially, it will just create the cache and download there if it
doesn't exist. There is no manual interaction that needs to take place
for any system or person that wants to use this.

Example of actions run first time and second time (runner knows nothing
about caching, no manual setup):
https://github.com/nod-ai/SHARK-TestSuite/actions/runs/8711532825
https://github.com/nod-ai/SHARK-TestSuite/actions/runs/8711798501

Final model testing job time: ~5 mins

The machine in the iree repo is even faster, so can expect lower times
:)

Fixes https://github.com/nod-ai/SHARK-TestSuite/issues/169