Add option to fail pytest collection if files are missing. (#180)
I wanted developers to be able to run the full test suite without
needing to download all files. That meant skipping tests if files were
not downloaded. On CI machines, we probably want to fail instead of
skipping tests, as we _should_ be downloading all files or pulling them
from a cache. This adds a flag that lets us control that behavior.