Add support so we can have a central location for spec files (#220)
This commit adds support so we can use environment variables in our
config files. This doesn't constrict us to only files in the directory.
I moved the spec file out to a central location. The `list2command` is
because when we use shell=true, 2nd and following arguments are for the
shell itself, so we need to use a string to pass a command to the shell.

Update: PR has been updated to show helpful, easy to reproduce error
messages:
Compile Failure Example:
https://github.com/nod-ai/SHARK-TestSuite/actions/runs/8945969568
Runtime Failure Example:
https://github.com/nod-ai/SHARK-TestSuite/actions/runs/8946009337