Pulling in Alt E2ESHARK CI changes (#388)
This pulls in the required changes to run nightly CI for alt e2eshark.

Few things to iterate on here:

- Figure out a way to run a loop/matrix on a step so we don't have a
bunch of download commands in push artifact step
- Figure out how to not fail the download artifact step if an artifact
for a certain shard is not available
- Pull the merge report script calls, creating directories, copy pasting
artifacts to eventually push into the reports repo into a python script
that won't fail if certain shards fail and are not available

Just merging this in so we have a nightly going that doesn't depend on
manual triggers. We can iterate on it. Has been running fine for the
last week.